test($RefereeTest): RefereeTest의 라운드 검증 테스트를 RoundExtractorTest로 분리
